webapp/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.0
pathspec==0.8.1
Pillow==8.3.0
psycopg2==2.9.1
python-dotenv==0.18.0
regex==2021.7.1
requests==2.25.1
toml==0.10.2
torch==1.9.0+cpu
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
